+++ FTHub/Model/MainApp/Settings/SettingsProfileDataNetworkModel.swift
@@ -30,7 +30,7 @@ struct SettingsProfileFormDataResponseData: NetworkCapable {
 struct SettingsProfileDataBody: NetworkCapable {
     let gender: String
     let age: Int
-    let height: Int
+    let height: Double
     let weight: Int
     let trainingFrequencyPerWeek: Int
     let units: String
